Fixes schmittjoh/twig.js#45 by escaping all . characters with backslashes
This is pain relief for the symptoms reported in the issue, but it doesn't cure
the underlying problem, which is that Javascript's string manipulation is
actually kind of sucky.

The ideal fix for this issue would have been to use Mozilla's non-standard third
String.replace parameter and replace the line with the following:

    str = str.replace(key, map[key], 'g');

This trivially fixes the whole problem, but it's useless because it only works
on Firefox.

Alternatively, it'd be nice if Javascript provided a native way of escaping
strings for use in regular expressions, like PHP's preg_quote
(http://php.net/preg_quote).

Since neither of these things are part of this reality, we are stuck with an
awkward problem. Perhaps a future version of twig.js ought to perform this
escaping for all such characters: . \ + * ? [ ^ ] $ ( ) { } = ! < > | : -

In the meantime, this is a nice safe, small commit, and it fixes the issue.