Introduce a Makefile to reduce configuration and setup, and streamline building
I now understand why this project uses Closure Compiler. However, I think the
recompilation step is a little clumsy and this is preventing people from
contributing to the project as effectively as they could be.

This commit removes the need for developers to go hunting for a copy of
plovr.jar and then configure $_SERVER['PLOVR_BIN'] in ./config.php. The Makefile
knows where to get plovr, installs it to a known location, and invokes it in
that location when building the assets.

It'll only attempt to fetch plovr if it's not already present at bin/plovr.
Anybody with edge-case needs who wants a specific version of plovr.jar can cope
by placing their version in bin/plovr, without impacting those who simply want
to get on with building the assets.